+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ## v0.0.3
 
+- rename `myClorm.HashableList` into `myClorm.ImmutableList`
 - introduce `multimap.Multimap[K,V]` as an immutable mappng from keys `K` to
   sets of values `V`
 - add custom converter hooks for generic aliases in myClorm